# docs/openapi/ferry.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ferry API",
    "description": "The HTTP API of [Ferry](https://github.com/Carter2307/ferry), a self-hosted Render alternative. The `ferry` CLI and the web dashboard use exactly this API.\n\n**Authentication.** Every `/api/v1` route needs the server's API token (printed by `ferryd` on first start, stored in `<data-dir>/api_token`) as `Authorization: Bearer <token>`; use the **Authorize** button to set it here. `GET` requests may pass it as `?access_token=<token>` instead, because browsers' `EventSource` can't set headers (prefer the header: query strings end up in logs). Webhooks (`/hooks/...`) authenticate with their own secrets, and `/healthz`, this document and Swagger UI need no token.\n\n**Conventions.** JSON bodies with snake_case fields; request bodies reject unknown fields and an empty body counts as `{}`. `{id}` path segments accept a resource id **or** its name. Errors are `ApiErrorBody` documents (`{\"error\": {\"code\", \"message\"}}`) with the matching HTTP status; unknown `/api` routes are JSON 404s. Boolean query flags accept `true/false`, `1/0`, `yes/no`, `on/off`, and an empty value means `true` (`?follow`).\n\n**Streams.** Logs and the change feed are Server-Sent Events (`text/event-stream`): see the `logs` operations and `GET /api/v1/events`. Idle streams get a `: keep-alive` comment every 15 s, and every stream ends when the server shuts down.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
...
  "paths": {
...
    "/api/v1/git/authorize": {
      "post": {
        "tags": [
          "git"
        ],
        "summary": "Authorize a git account in the browser",
        "description": "Starts connecting an account on the provider's own pages and answers with where to send the browser (`status: redirect`): a URL to navigate to (`method: get`) or to submit a form with `fields` to (`method: post`). When the provider is done it sends the browser back to `redirect_uri` (the dashboard's `/git/callback` page) with query parameters to hand to `POST /api/v1/git/callback`.\n\n**GitHub**: the browser posts a manifest to GitHub, which registers a private GitHub App for this server (on the user's account, or in `organization`) and then asks the account which repositories the app may read. No token is ever typed: the app's key mints short-lived tokens.\n\n**GitLab**: the account authorizes an OAuth application created for this server (scopes `read_api` and `read_repository`, redirect URI = `redirect_uri`). Give its `client_id` and `client_secret` the first time (400 `git_application_required` otherwise); they are kept for later authorizations.\n\nWith `connection_id`, resumes a `pending` connection or authorizes a connection again; the answer is `status: connected` when nothing is left to do (a GitHub App that is already installed).",
        "operationId": "authorizeGit",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorizeGit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Where to send the browser, or the connection when it is already complete.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitAuthorization"
                }
              }
            }
          },
          "400": {
            "description": "Invalid `redirect_uri`, `base_url` or organization; a provider that needs its OAuth application first (code `git_application_required`); or the provider refused (code `git_authorization_rejected`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such git connection (`connection_id`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "502": {
            "description": "The provider could not be reached, or answered an error (code `git_provider_unavailable`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error (`docker_error` as 502, `database_error`, `internal_error`...).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/git/branches": {
      "get": {
        "tags": [
          "git"
        ],
        "summary": "List the branches of a repository",
        "description": "The branches of a repository and its default one, asked from the repository's remote itself (`git ls-remote`) on every call. Works for any repository URL a service can deploy from; it is read the way a deploy would clone it: with the git connection that serves the URL when there is one (`connection_id` in the answer), with the credentials the URL carries, or without any.",
        "operationId": "listGitBranches",
        "parameters": [
          {
            "name": "repo_url",
            "in": "query",
            "description": "The repository: any URL a service can deploy from (`https://…`,\n`git@host:owner/name.git`, a path on the server).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The branches.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitBranches"
                }
              }
            }
          },
          "400": {
            "description": "Malformed repository URL.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "502": {
            "description": "The remote refused, wasn't found or didn't answer in time (code `git_remote_unreachable`): the message says why, and what to do when it looks like a private repository.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error (`docker_error` as 502, `database_error`, `internal_error`...).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/git/callback": {
      "post": {
        "tags": [
          "git"
        ],
        "summary": "Finish a step of a browser authorization",
        "description": "Takes the query parameters the provider sent the browser back with (to the `redirect_uri` given to `POST /api/v1/git/authorize`): `state`, and `code` and / or `installation_id`. Answers with the next page to send the browser to (`status: redirect`: after GitHub registered the app, its installation page) or with the connection (`status: connected`). A `state` works once, for an hour. Without `state` but with an `installation_id` (GitHub sends the browser back after an installation was changed on its own pages), the connection of that installation is read again.",
        "operationId": "gitCallback",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GitCallback"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The next step, or the connection.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitAuthorization"
                }
              }
            }
          },
          "400": {
            "description": "Unknown, expired or already used `state`; a missing parameter; or the provider refused: the user denied the authorization, the code is no longer valid (code `git_authorization_rejected`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "502": {
            "description": "The provider could not be reached, or answered an error (code `git_provider_unavailable`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error (`docker_error` as 502, `database_error`, `internal_error`...).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GitAuthorization": {
        "type": "object",
        "description": "The next step of a browser authorization (`POST /api/v1/git/authorize`\nand `POST /api/v1/git/callback`).",
        "required": [
          "status"
        ],
        "properties": {
          "connection": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/GitConnectionView",
                "description": "`connected`: the connection."
              },
              {
                "type": "null"
              }
            ]
          },
          "fields": {
            "type": "object",
            "description": "`redirect` with `post`: the fields of the form to submit.",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "method": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/GitRedirectMethod",
                "description": "`redirect`: how."
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/GitAuthorizationStatus"
          },
          "url": {
            "type": [
              "string",
              "null"
            ],
            "description": "`redirect`: the provider's page to send the browser to."
          }
        }
      },
...
      "GitRedirectMethod": {
        "type": "string",
        "description": "How the browser is sent to the provider.",
        "enum": [
          "get",
          "post"
        ]
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API token",
        "description": "The server's API token (`ferryd` prints it on first start; it is stored in `<data-dir>/api_token`). `GET` requests may send it as `?access_token=` instead."
      }
    }
  },
...
}